refactor: setup the Kitsu:Library to work within the app layer
Update the application layer to correctly inject the new Kitsu:Library classes
