+++ app/build.gradle
@@ -82,12 +82,13 @@ dependencies {
     implementation project(path: ":libraries:encryption")
     implementation project(path: ":libraries:kitsu")
     implementation project(path: ":libraries:kitsu:auth")
+    implementation project(path: ":libraries:kitsu:library")
     implementation project(path: ":libraries:kitsu:search")
     implementation project(path: ":libraries:kitsu:trending")
     implementation project(path: ":libraries:kitsu:user")
     implementation project(path: ":libraries:library")
-    implementation project(path: ":libraries:server")
     implementation project(path: ":libraries:server:auth")
+    implementation project(path: ":libraries:server:library")
     implementation project(path: ":libraries:server:search")
     implementation project(path: ":libraries:server:trending")
     implementation project(path: ":libraries:server:user")
