+++ features/series/build.gradle
@@ -39,7 +39,6 @@ dependencies {
     implementation project(path: ":libraries:account")
     implementation project(path: ":libraries:core")
     implementation project(path: ":libraries:library")
-    implementation project(path: ":libraries:server")
 
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
     implementation "androidx.appcompat:appcompat:$appcompat_version"
