+++ libraries/account/build.gradle
@@ -35,7 +35,6 @@ dependencies {
 
     implementation project(path: ":libraries:core")
     implementation project(path: ":libraries:database")
-    implementation project(path: ":libraries:server")
     implementation project(path: ":libraries:server:user")
 
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
