+++ libraries/library/build.gradle
@@ -35,7 +35,7 @@ dependencies {
 
     implementation project(path: ":libraries:core")
     implementation project(path: ":libraries:database")
-    implementation project(path: ":libraries:server")
+    implementation project(path: ":libraries:server:library")
 
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
     implementation "androidx.appcompat:appcompat:$appcompat_version"
